+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/Imp.hs
@@ -4,6 +4,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Cardano.Ledger.Babbage.Imp (spec) where
 
@@ -13,6 +14,7 @@ import Cardano.Ledger.Alonzo.Rules (
   AlonzoUtxosPredFailure,
   AlonzoUtxowPredFailure,
  )
+import Cardano.Ledger.Babbage (BabbageEra)
 import Cardano.Ledger.Babbage.Core (BabbageEraTxBody, InjectRuleFailure)
 import Cardano.Ledger.Babbage.Rules (BabbageUtxoPredFailure, BabbageUtxowPredFailure)
 import Cardano.Ledger.Babbage.TxInfo (BabbageContextError)
@@ -33,6 +35,7 @@ import Test.Cardano.Ledger.Imp.Common
 spec ::
   forall era.
   ( AlonzoEraImp era
+  , EraSpecificSpec era
   , BabbageEraTxBody era
   , InjectRuleFailure "LEDGER" ShelleyDelegPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyPoolPredFailure era
@@ -49,7 +52,10 @@ spec ::
 spec = do
   AlonzoImp.spec @era
   withEachEraVersion @era $
-    describe "BabbageImpSpec" $ do
+    describe "BabbageImpSpec - era generic tests" $ do
       Utxo.spec
       Utxow.spec
       Utxos.spec @era
+
+instance EraSpecificSpec BabbageEra where
+  eraSpecific = AlonzoImp.alonzoEraSpecificSpec