+++ eras/dijkstra/testlib/Test/Cardano/Ledger/Dijkstra/Imp.hs
@@ -4,6 +4,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Cardano.Ledger.Dijkstra.Imp where
 
@@ -14,20 +15,19 @@ import Cardano.Ledger.Babbage.TxInfo
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Conway.Rules
 import Cardano.Ledger.Conway.TxInfo
+import Cardano.Ledger.Dijkstra (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Core
-import Cardano.Ledger.Plutus
 import Cardano.Ledger.Shelley.API
 import Cardano.Ledger.Shelley.Rules
 import Data.Typeable (Typeable)
 import Test.Cardano.Ledger.Common
-import Test.Cardano.Ledger.Conway.Imp as ConwayImp
+import qualified Test.Cardano.Ledger.Conway.Imp as ConwayImp
 import Test.Cardano.Ledger.Dijkstra.ImpTest
-import Test.Cardano.Ledger.Imp.Common
 
 spec ::
   forall era.
   ( DijkstraEraImp era
-  , EraPlutusTxInfo 'PlutusV2 era
+  , EraSpecificSpec era
   , Inject (BabbageContextError era) (ContextError era)
   , Inject (ConwayContextError era) (ContextError era)
   , InjectRuleFailure "LEDGER" ConwayGovPredFailure era
@@ -37,7 +37,6 @@ spec ::
   , InjectRuleFailure "LEDGER" AlonzoUtxoPredFailure era
   , InjectRuleFailure "LEDGER" AlonzoUtxosPredFailure era
   , InjectRuleFailure "LEDGER" AlonzoUtxowPredFailure era
-  , InjectRuleFailure "LEDGER" ShelleyDelegPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxoPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure era
   , InjectRuleFailure "LEDGER" ConwayDelegPredFailure era
@@ -54,10 +53,9 @@ spec ::
   , ToExpr (Event (EraRule "ENACT" era))
   , Eq (Event (EraRule "ENACT" era))
   , Typeable (Event (EraRule "ENACT" era))
+  , ToExpr (Event (EraRule "BBODY" era))
   ) =>
   Spec
-spec = do
-  ConwayImp.spec @era
-  withImpInit @(LedgerSpec era) $ do
-    -- TODO
-    pure ()
+spec = ConwayImp.spec @era
+
+instance EraSpecificSpec DijkstraEra