+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Imp.hs
@@ -3,10 +3,12 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Cardano.Ledger.Shelley.Imp (spec) where
 
 import Cardano.Ledger.Core
+import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.Rules (
   ShelleyPoolPredFailure,
   ShelleyUtxoPredFailure,
@@ -24,12 +26,14 @@ import qualified Test.Cardano.Ledger.Shelley.UnitTests.InstantStakeTest as Insta
 spec ::
   forall era.
   ( ShelleyEraImp era
+  , EraSpecificSpec era
   , InjectRuleFailure "LEDGER" ShelleyUtxoPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyPoolPredFailure era
   ) =>
   Spec
 spec = do
+  describe "Era specific tests" . withEachEraVersion @era $ eraSpecific
   describe "ShelleyImpSpec" $ withEachEraVersion @era $ do
     Epoch.spec
     Ledger.spec
@@ -38,3 +42,5 @@ spec = do
     Utxo.spec
   describe "ShelleyPureTests" $ do
     Instant.spec @era
+
+instance EraSpecificSpec ShelleyEra