+++ eras/conway/impl/CHANGELOG.md
@@ -96,6 +96,7 @@
 
 ### `testlib`
 
+* Added `EraSpecificSpec ConwayEra` instance
 * Added `registerRewardAccountWithDeposit`
 * Added `regDelegToDRep`
 * Generalised the following helpers and thus changed their constraints to `ConwayEraImp`: